backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
joblib
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-dotenv
python-multipart